Include immediate and refined distance estimates if available when collecting raw samples.

+++ app/src/main/java/com/jifalops/localization/datatypes/RangingParams.java
@@ -128,7 +128,7 @@ public class RangingParams {
 
 
     public interface Sample {
-        int[] getInputs();
+        double[] getInputs();
     }
